different way of getting good quality vid embed
I only need the video to be 360x180, but the 180p encoding is too
blurry, so I have to cram the 360p encoding in there instead.

Making the iframe just a tad bit larger than 360x180 did the trick, but
it looks like another solution is to use the parameter vq=medium.  Which
is... sort of undocumented? But maybe better than using an odd iframe
size.  We'll see.

(It would be nice if I could just delete the 180p encode from YouTube.)