respond to MegaPath losing my file hosting server
For now, change all the download links to explicitly point to the (temporary?) mirror. I should also change the DNS for files.neogeographica.com to point there in case a link using that name has escaped into the wild… then I can change the links back to using that name.